This cleans up the "easy" warning fixes which can be made using my Regress script, comparing object-files before/after the edits:
	https://invisible-island.net/ansification/index.html
	https://invisible-island.net/scripts/readme.html

The changes are casts, which quiet the gcc warnings about implicit
conversion that my "gcc-normal" script would show.  I avoided
reformatting the code.

The change reduces the number of gcc warnings from 769 to 163.

+++ src/Core.c
@@ -349,7 +349,7 @@ static Boolean CoreSetValues(
 	   }
 	   else {
 	       attributes.background_pixmap = new->core.background_pixmap;
-	       window_mask &= ~CWBackPixel;
+	       window_mask &= (unsigned long) (~CWBackPixel);
 	       window_mask |= CWBackPixmap;
 	   }
 	   redisplay = TRUE;
@@ -366,7 +366,7 @@ static Boolean CoreSetValues(
 	   }
 	   else {
 	       attributes.border_pixmap = new->core.border_pixmap;
-	       window_mask &= ~CWBorderPixel;
+	       window_mask &= (unsigned long) (~CWBorderPixel);
 	       window_mask |= CWBorderPixmap;
 	   }
        }
